物流敘述改一下，新增訂單的驗證改成抓db

+++ models/frontend/platform_logistics.go
@@ -1,17 +1,17 @@
 package frontend
 
 type PlatformLogistics struct {
-	PlatformID      int  `json:"-"`
-	HomeEnabled     bool `json:"home_enabled"`
-	UniEnabled      bool `json:"uni_enabled"`
-	FamilyEnabled   bool `json:"family_enabled"`
-	HilifeEnabled   bool `json:"hilife_enabled"`
-	OkEnabled       bool `json:"ok_enabled"`
-	HomeChargeFee   int  `json:"home_charge_fee"`
-	UniChargeFee    int  `json:"uni_charge_fee"`
-	FamilyChargeFee int  `json:"family_charge_fee"`
-	HilifeChargeFee int  `json:"hilife_charge_fee"`
-	OkChargeFee     int  `json:"ok_charge_fee"`
+	PlatformID      int     `json:"-"`
+	HomeEnabled     bool    `json:"home_enabled"`
+	UniEnabled      bool    `json:"uni_enabled"`
+	FamilyEnabled   bool    `json:"family_enabled"`
+	HilifeEnabled   bool    `json:"hilife_enabled"`
+	OkEnabled       bool    `json:"ok_enabled"`
+	HomeChargeFee   float32 `json:"home_charge_fee"`
+	UniChargeFee    float32 `json:"uni_charge_fee"`
+	FamilyChargeFee float32 `json:"family_charge_fee"`
+	HilifeChargeFee float32 `json:"hilife_charge_fee"`
+	OkChargeFee     float32 `json:"ok_charge_fee"`
 	TimeDefault
 }
